Template audit.rules, toggle immutability
TL;DR
=====
- Allow immutability to be toggled with a boolean
- Template the entire `audit.rules` file
- Allow customization of the `audit.rules` path

Detail
======
The audit daemon (auditd) is equipped with a set of rules that are
pre-baked into stemcells. The prior auditd job within this repo simply
appended to those rules, and didn't expose an option to modify the
rules. We decided to take the approach to template the entire file, so
we can easily conform to the Ubuntu 16.04 DISA STIG.

Immutability is technically a rule, but having a boolean to control it
seemed like a much cleaner approach than forcing the operator to use it.
If an operator made a typo in their auditd rules, they would ultimately
have to recreate the VM to continue testing; pretty bad for a fast
feedback loop.

The auditd spec now exposes `immutable`, a boolean for controlling
immutability and `rules_path`, a string that allows you to define an
alternate rules path.

The rules are now correctly spaced when being written, because we make
use of the `-` (trim) operator.